Hopefully adapted grid scaling
Should now behave reasonably and should make it possible to make sense
of the graph even in long term situation. Hopefully this was the last
thing for issue #5 and we can close it. Will push to openrepos for
testing and after a while to harbour as well unless there will be some
issues discovered during testing.